Implementation of the class Common class that received all the methods from the class HomePage and i ade all the other page herit from the class CommonClass
