+++ src/main/java/ContactPage.java
@@ -1,10 +1,9 @@
 import org.openqa.selenium.WebDriver;
 
-public class ContactPage {
-    WebDriver driver;
+public class ContactPage extends CommonClass {
 
     public ContactPage(WebDriver driver){
-        this.driver = driver;
+        super(driver);
     }
 
     public ProductListPage SearchByAutoComplete(){
